added new datepicker to invoices, new client status feature, removed delete from client for historical

+++ app/Http/Controllers/ServiceController.php
@@ -103,6 +103,13 @@ class ServiceController extends Controller
      */
     public function destroy(Request $request, Service $service)
     {
+        // $service = Service::find($id);
+
+        if (!$service) {
+            return redirect('/invoice')->with('alert', 'Invoice Not Found');
+        }
+
+        $service->invoices()->detach();
         $service->delete();
     }
 }
